Add x-velocity variety based on paddle impact
Pretty pleased with this change, it adds a bit of life to the gameplay which wasn't there before. The ball.dx changes by a varying degree based on where on the paddle it hits! Pretty cool, huh? There's 5 sections of the paddle. The left and right most portions increase the x velocity the most, and the center portion doesn't change it at all. Adds a degree of skill.