Add side-of-brick collision detection
If a ball hits the side of a brick it will (9/10 times...) reverse its x direction. Kind of a painful process, but it became evident that it is crucial that the gameloop check for side of brick collection BEFORE it checked for collision in general. Deleting the brick from a side hit was negating the desired effect.